avanço++

+++ Assets/Scripts/HealthBar.cs
@@ -16,7 +16,7 @@ public class HealthBar : MonoBehaviour
         healthBar.value = playerHealth.maxHealth;
     }
 
-    public void SetHealth(int hp)
+    public void SetHealth(float hp)
     {
         healthBar.value = hp;
     }
